Add alternate layout functionality
