+++ ishell.py
@@ -1,0 +1,2 @@
+from IPython import embed
+embed()
